python-worker/requirements.txt
redis
google-generativeai